added switch tokens

+++ app/src/main/java/com/numtory/application/features/gold/data/models/GoldikaDataModel.kt
@@ -23,9 +23,9 @@ data class GoldikaDataDataModel(
 
 data class GoldikaPriceDataModel(
     @SerializedName("sell")
-    val sell: Int,
+    val sell: Long,
     @SerializedName("buy")
-    val buy: Int,
+    val buy: Long,
 ) {
     fun toEntity(): GoldikaPrice = GoldikaPrice(sell, buy)
 }
