+++ app/src/main/java/com/numtory/application/features/gold/data/models/TlynDataModel.kt
@@ -29,9 +29,9 @@ data class TlynDataDataModel(
 
 data class TlynPriceDataModel(
     @SerializedName("sell")
-    val sell: Int,
+    val sell: Long,
     @SerializedName("buy")
-    val buy: Int,
+    val buy: Long,
 ) {
     fun toEntity(): TlynPrice = TlynPrice(sell, buy, )
 }
