+++ app/src/main/java/com/numtory/application/features/gold/domain/entities/Tlyn.kt
@@ -15,7 +15,7 @@ data class TlynData(
 }
 
 data class TlynPrice(
-    val sell: Int,
-    val buy: Int,
+    val sell: Long,
+    val buy: Long,
 ) {
 }
