+++ app/src/main/java/com/numtory/application/features/market/domain/usecase/RemoveInvalidExchangeUseCase.kt
@@ -9,8 +9,8 @@ class RemoveInvalidExchangeUseCase constructor() {
 
         return params.markets
             .filter {
-                ((it.sellPrice != null && it.sellPrice?.toFloat() != 0f) && (it.buyPrice != null && it.buyPrice?.toFloat() != 0f))
-                        || it.marketPrice != null && it.marketPrice?.toFloat() != 0f
+                ((it.sellPrice != null && it.sellPrice?.toDouble() != 0.0) && (it.buyPrice != null && it.buyPrice?.toDouble() != 0.0))
+                        || it.marketPrice != null && it.marketPrice?.toDouble() != 0.0
             }
 //            .filter { marketPrice ->
 //                if (params.exchangesInfo?.isNotEmpty() != true) true
